Move build_qurl to picard.util.qt
build_qurl constructs a QUrl, so it moves to the Qt-only picard.util.qt.
Update the importers (oauth, util.mbserver, browser.filelookup).
