+++ picard/util/__init__.py
@@ -87,7 +87,6 @@ from picard import (
     tagger_instance,
 )
 from picard.const import (
-    MUSICBRAINZ_SERVERS,
     PICARD_DOCS_URLS,
     PICARD_URLS,
 )
@@ -741,37 +740,6 @@ def get_url(url_key: str) -> str:
 
     # No match in defined Picard URLs
     return url_key
-
-
-def build_qurl(
-    host: str, port: int = 80, path: str | None = None, queryargs: Mapping[str, Any] | None = None
-) -> QtCore.QUrl:
-    """
-    Builds and returns a QUrl object from `host`, `port` and `path` and
-    automatically enables HTTPS if necessary.
-
-    Encoded query arguments can be provided in `queryargs`, a
-    dictionary mapping field names to values.
-    """
-    url = QtCore.QUrl()
-    url.setHost(host)
-
-    if port == 443 or host in MUSICBRAINZ_SERVERS:
-        url.setScheme('https')
-    elif port == 80:
-        url.setScheme('http')
-    else:
-        url.setScheme('http')
-        url.setPort(port)
-
-    if path is not None:
-        url.setPath(path)
-    if queryargs is not None:
-        url_query = QtCore.QUrlQuery()
-        for k, v in queryargs.items():
-            url_query.addQueryItem(k, str(v))
-        url.setQuery(url_query)
-    return url
 
 
 def union_sorted_lists(list1: Sequence, list2: Sequence) -> list:
